CoPilot agent (#84)
* adding initial commit for the copilot poc

* further fixes

* added new engine for the copilot apis

* wip

* - added the new vm info to copilot
- added the new info to the remote hosts and catalogs

* fixes to the build system

* final changes to the new version

* fix format

* fix issue with the pipeline

* adding missing vscode.d.ts to avoid errors

* fix issues width lint

+++ src/tree/commands/devopsRemoteHostProvider/removeVirtualMachine.ts
@@ -48,16 +48,28 @@ const registerDevOpRemoveVirtualMachineCommand = (
             vscode.window.showErrorMessage(`Machine ${item.name} not found`);
             return;
           }
+          if (machine.State === "invalid") {
+            const ok = await DevOpsService.unregisterRemoteHostVm(provider, machineId).catch(reject => {
+              vscode.window.showErrorMessage(`${reject}`);
+              foundError = true;
+              return;
+            });
 
-          const ok = await DevOpsService.removeRemoteHostVm(provider, machineId).catch(reject => {
-            vscode.window.showErrorMessage(`${reject}`);
-            foundError = true;
-            return;
-          });
+            if (!ok || foundError) {
+              vscode.window.showErrorMessage(`Failed to remove virtual machine ${item.name}`);
+              return;
+            }
+          } else {
+            const ok = await DevOpsService.removeRemoteHostVm(provider, machineId).catch(reject => {
+              vscode.window.showErrorMessage(`${reject}`);
+              foundError = true;
+              return;
+            });
 
-          if (!ok || foundError) {
-            vscode.window.showErrorMessage(`Failed to remove virtual machine ${item.name}`);
-            return;
+            if (!ok || foundError) {
+              vscode.window.showErrorMessage(`Failed to remove virtual machine ${item.name}`);
+              return;
+            }
           }
 
           DevOpsService.refreshRemoteHostProviders(true);
